Split API and CLI modules. Relates to #29
