+++ layrry-launcher/pom.xml
@@ -33,11 +33,7 @@
   <dependencies>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>layrry-config-yaml</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>${project.groupId}</groupId>
-      <artifactId>layrry-config-toml</artifactId>
+      <artifactId>layrry-config</artifactId>
     </dependency>
     <dependency>
       <groupId>org.slf4j</groupId>
@@ -52,11 +48,6 @@
       <artifactId>shrinkwrap-resolver-api-maven</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.beust</groupId>
-      <artifactId>jcommander</artifactId>
-      <version>1.78</version>
-    </dependency>
-    <dependency>
       <groupId>io.methvin</groupId>
       <artifactId>directory-watcher</artifactId>
       <version>0.9.9</version>
@@ -69,34 +60,4 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-shade-plugin</artifactId>
-        <version>3.2.4</version>
-        <configuration>
-          <finalName>layrry-cli-${project.version}</finalName>
-          <transformers>
-            <transformer
-                    implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-              <manifestEntries>
-                <Main-Class>org.moditect.layrry.Layrry</Main-Class>
-              </manifestEntries>
-            </transformer>
-            <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
-          </transformers>
-        </configuration>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
 </project>
