+++ vertx-example/layrry-links-runner/pom.xml
@@ -34,7 +34,7 @@
   <dependencies>
     <dependency>
       <groupId>org.moditect.layrry</groupId>
-      <artifactId>layrry-launcher</artifactId>
+      <artifactId>layrry-cli</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -67,7 +67,7 @@
             <arguments>
               <argument>-classpath</argument>
               <classpath/>
-              <argument>org.moditect.layrry.Layrry</argument>
+              <argument>org.moditect.layrry.cli.LayrryRunner</argument>
               <argument>--layers-config</argument>
               <argument>src/test/resources/layers.yml</argument>
             </arguments>
